Dobble use of one word, that that.
I do have some sort of OCD, and now I have read this for something, I
like to fix it :-)

Under WaitForMSIExec, the word That is dobble used.

Ord text:
WaitForMsiExec
Sometimes an EXE bootstrapper will launch an MSI install. In such cases,
this variable will ensure that that this function waits for the msiexec
engine to become available before starting the install.

I have removed one of the that, se below:
ensure that this function.