Add  Function  Get-ScheduledTask_EPAU
Add  Get-ScheduledTask_EPAU function and reference it from Function execute-processasuser. The goal is to remove the English text as non English OS will just fail running that function with -wait. The task scheduler service is returning a value instead of text. It was test on Windows 8.1 French and Windows 7 Japanese.