Proposed Fix for Issue #335
Proposed Fix for Issue #335 and in extension #200.

Change the use of Environment.GetCommandLineArgs to Environment.CommandLine and then manually parse it in the script. We'll need to split the string on the hyphen character as this is the character used to denote arguments in PoweShell. This will cause an issue with strings that contain hyphens such as "Deploy-Application", but if we split on a space and a hyphens (ex " -") as all arguments have to be separated by spaces preceding the hyphens. Line 84/85 has to be modified as the split command I'm using removes the hyphen so I have to put it back in somewhere.

A more ideal solution would be to use Regex to do the splitting to preserve the hyphen but that requires an additional reference not currently being used, but I believe this subject should be discussed further.