Execute-Process -MsiExecWaitTime parameter fix
The parameter -MsiExecWaitTime is creating a timespan variable through New-TimeSpan when its not specified but when it is, its created through the variable which creates a timespan variable with incorrect value set, because it assumes you want to use that value for Ticks not seconds. This change fixes it by changing it to an integer and then creating a new timespan variable where this integer is used for seconds.