Add function Remove-InvalidFileNameChars
This patch adds the function Remove-InvalidFileNameChars to the toolkit and is used in places where filename was being validated before. I also added removal of invalid chars to the New-Zipfile command since it doesn't have one, causing it to fail while creating zip files when an invalid character is specified.