Get-InstalledApplication - add more log messages
When this function looks for the specified application, it will skip entries that look similar to updates. However, the user doesn't get any information about this happening. This pull request fixes it by writing the number of such entries to the log.

We could also list all the entries that are skipped however such list could be very long.

This pull request also adds a log entry when no result is found as that is currently missing.