BREAKING CHANGE: Modify Execute-Process -ContinueOnError behaviour
Currently Execute-Process is one of the functions that fails on error even if ContinueOnError is specified because it is used for ignoring exit codes and not handling errors. This PR fixes it by adding the ability to ignore all exit codes through -IgnoreExitCodes "*" while -ContinueOnError will be used for ignoring errors that occured inside the function as expected.