BREAKING CHANGE: Modify Execute-Process -ContinueOnError behavior
Currently Execute-Process is one of the functions that fails on error even if ContinueOnError is specified because this parameter is used for ignoring exit codes and not handling errors. This PR fixes it by adding the ability to ignore all exit codes through -IgnoreExitCodes "*" while -ContinueOnError will be used for ignoring errors that occurred inside the function as expected.

Also fixes a couple of typos and extends description for -PassThru because currently, it makes the function ignore exit codes and that is not documented. 

Rewords log entry for when this function is used for an MSI installation but another installation is already running.

#125