Add -UseShellExecute for Execute-Process
WindowStyle doesnt usually work without this parameter on. CreateNoWindow will usually not work with this on.

This parameter changes the way the application is started. It is similar to entering a path into the Run dialog or into the Windows Explorer. If it is a file not an application, then the Shell will open it in a program assigned for this extension. The Shell doesnt stand for command shell but for operating system shell. It changes how Working Directory is used.

Expanded description for WindowStyle,CreateNoWindow and WorkingDirectory

If this parameter is set to $true there is no output/error stream. Only an exit code.