Add config paths without admin rights
Adds config paths for installations where RequireAdmin is False.

Also changes default -Source value to a more useful value: it will be set to Parent's function name and if not inside a function then the name of the script. If not inside a function or a script, then the value will be set to 'Unknown'