Add a return value if NoWait and PassThru are specified
Allows the function to retun an object if PassThru is specified and NoWait is specified also. Normally we dont have an output since the process hasnt stopped yet but instead of returning nothing we can retun process id, process name and process handle if a user wants to check something through them.