Rename Restart Later to Minimize
This button is misleading for the user. Microsoft uses Restart Later buttons to extend the duration of time left before a reboot is triggered. The function Show-InstallationRestartPrompt however does not extend the duration even though the button is called Restart Later. It only minimizes the window. This PR renames the button to Minimize.