TempPath change
Since $envTemp points to c:\windows\temp and the folder has limited permissions for Administrators by default, it has to be changed again. ProgramData seems like the best option other than ProgramFiles which is used for programs themselves. ProgramData is also hidden by default.

ProgramData is read only to users by default so we need to change the log dir to a user writable location for when the toolkit runs without admin rights. Not many such folders exist. One of the best locations is LocalAppData since those folders dont get erased without reason and are user writable.

Also modifies Block and Unblock AppExecution functions so they get bypassed if admin rights are not requested because the functions require elevated permissions to function.