Clean a couple of global variables on exit
If you run multiple toolkits in the same powershell session, these variable do not get reinitialized because they are not empty. This causes bugs, for example the log file is the same even though the toolkit is not. This resolves it but not if the script is exited without Exit-Script.