Change default log location based on deployment system
Check if script is executing from ConfigMgr or Intune and if so, change the default log location so that logs can be collected via diagnostics data collection of those respective systems. This option is enabled by default but an additional XML config option has been added to disable this functionality. Proposing to include this as default with the 4.0 release as it can be considered a breaking change.