- Make `Invoke-ADTAllUsersRegistryChange` work on a scriptblock again.
The problem here is we don't really know what the caller was doing in the scriptblocks, and could have been doing anything beyond the given examples in the comment-based help.

This required restoring `Invoke-ScriptBlockInSessionState` and parsing dynamically recompiling the scriptblock to make it work right, but it does the job and ensures 1:1 backwards compatibility.