- Rework `Invoke-HKCURegistrySettingsForAllUsers` into `Invoke-ADTAllUsersRegistryChange`.
The interface absolutely needed to change here. The previous setup depending on a scriptblock being passed through required detailed knowledge of how the function works internally which we can't have.

Further to this, the setup a complete no-op when it comes to being a module due to the incoming scriptblock being bounded to the caller's location. Setups like `$RegistrySettings.Ast.GetScriptBlock()` do generate an unbound scriptblock, but if the scriptblock had variables in the caller's scope, they'd be inaccessible. Conversely, abusing `ForEach-Object` with a setup like `ForEach-Object -Process $RegistrySettings` would allow the scriptblock to invoke in its bound location, but then it wouldn't be able to access `$UserProfile` within the function.