- Abuse `ForEach-Object` in `Invoke-ADTAllUsersRegistryChange` so we don't depend on the caller's session state.
Using the caller's session state here might be incorrect as there's nothing stopping anyone from invoking `Invoke-ADTAllUsersRegistryChange` in a sub-function, which would have its own session state.

We also cannot redefine the scriptblock internally and call it directly as it would allow the caller to read all of the module parameters and break security.

We also cannot leverage `-PipelineVariable` on a pipelined command before `ForEach-Object` as the module has no access to the caller's scope to be able to define `$UserProfile` within the scriptblock.

As such, we invoke the scriptblock with `ForEach-Object`, and pass `$UserProfile` through as the piped object. This change does mean that moving forward for new developments, people will need to use `$_` or `$PSItem` instead of the old variable, but its well worth it in the end. If callers are doing loops within the scriptblock, it'll be up to them to archive the value off before doing so.