- Robustness fix for `Show-ADTInstallationProgress`.
The `$Error` array could sometimes be added to the sync hash, even though it was empty. Because we were simply checking for the `Error` key in the sync hash and not its count, this caused erroneous failures.

Cover this both ways, only adding `$Error` to the sync hash if there's errors, and then only processing errors in the sync hash if the value isn't empty.