- Set all WinForms defaults in the module's psm1 code.
It's supposed to be set in your application as one of the first things you do, so setting it here is the best place.

We've defined `[System.Windows.Forms.Application]::SetCompatibleTextRenderingDefault($false)` globally within the module but have left it commented out as none of the form padding is catering for it. This has been there since .NET Framework 2.0 and uses the newer TextRenderer class instead of the old GDI+ interface. This helps make text render better, is more performant, and is better with internationalisation; and probably should have always been enabled from the get-go. See https://learn.microsoft.com/en-us/dotnet/api/system.windows.forms.application.setcompatibletextrenderingdefault?view=netframework-2.0#remarks for more information.

WinForms is on the chopping block anyway, so it's not super important. If WinForms were to remain, I'd want to enable this and re-work the padding on all form controls to accommodate.