- Refactor `PSAppDeployToolkit` module into submodules.
The monolithic design was what we needed when we were starting out, however its structurally better to start breaking things down. As we also need to support the classic dialogs in addition to the modern replacements, each requiring global states, it makes sense that they maintain their state within their own module.

It's also our intention that extensions to the project be delivered as modules. If we're shipping our own modules built on `PSAppDeployToolkit.Core` and `PSAppDeployToolkit.Common`, so can the community.