- Move `[ADTProcessObject]` class into the C# code as `[PSADT.Types.ProcessObject]`.
We're really hitting the limits of classes with PowerShell here, especially when it comes to classes that need to be used across different modules, etc.

Some consideration should be given towards `[ADTSession]` as well, but that's a much taller order since it needs to call on functions within `PSAppDeployToolkit.Core`. Such functionality would require re-writing in C# to make everything work.