- Ensure `-ComputerName` is removed from all Service functions and that wrappers handle this.
The parameter was half removed, but not entirely. `Get-Service` on PowerShell 7 simply doesn't support specifying a computer name anymore, so neither do we.

The requirements of doing a software task on one PC that's predicated on the service state of another should be next to non-existent, so I don't see any problem with this.