- Extend the callback system out with starting and finishing callback hooks.
The callbacks we were defining as closing were only executing for the final session, not on the close of every session; however the callbacks we were defining as opening were executing upon the opening of every session.

By extending this out fully with starting and finishing callback hooks, we can correctly place all callbacks where they should be called.