- Consolidate the Dialogs module back into the main PSAppDeployToolkit module.
These were initially separated out due to some specific requirements necessitating it at the time. It's no longer necessary and for the simplicity of the project, where a user expects that installing `PSAppDeployToolkit` gives them everything, it makes sense to bring it all back in.

The separation was a valuable experience as it assisted in architecting the module in a way that is compatible with supplemental modules, which is how we envisage the community will build upon and extend PSAppDeployToolkit in the future.

This consolidation process also ensures that our CommandTable is used for all PowerShell cmdlets and functions, ensures that there is no collision between dialog data, and leverages a dispatcher table to accelerate underlying dialog lookups.