- Adapt `PSAppDeployToolkit.psm1` and build system to our requirements.
It should be noted that Catesta removes the private and public folders, but leaves the Classes folder there. This is likely a bug, however we must exploit this as we can't have a class in the psm1 directly that relies on structs or classes within our C# code as classes are checked at compile time, not at runtime (such as when `Add-Type` is called).

Once we have a compiled C# file that can be loaded with the module manifest's `RequiredAssemblies` array, none of this will be a problem.