- Add FilterScript support to `Get-ADTInstalledApplication`.
This cleans up a lot of code and bad setup, and provides more options to the caller. The compat wrapper has also been amended.

Please note the deliberate use/abuse of `ForEach-Object` to invoke the scriptblock delegate provided by the caller. This allows the scriptblock to be invoked within its bound scope without recreating it inside the function, potentially exposing the internals of the function or module.

I considered the use of delay-binding here, however it only works with functions accepting pipelined input, and creating an internal sub-function just for this when `ForEach-Object` can be used in the manner committed is over-complication that simply isn't necessary.