- Remove all hard-coded invocations via $Script:CommandTable in prep for automating this via the module build process.
There's still a few usages within the code, but these are function pointers used with `Invoke-ADTCommandWithRetries` or as callbacks. These need to remain as `Invoke-ADTCommandWithRetries` is only going to look up a function via the CommandTable anyway, and the callback code expects function pointers at all times.

Lastly, this does not touch the CommandTable usage within the module init code as that is of no consequence to mocking and is required to ensure the security of the project.