- Remove padding on all hashtable definitions.
It's not a good idea as when a longer key needs adding, all other key/value pairs need adjustment. This creates noise in the commits and complicates bisection due to all the git blame noise. See "Add QuietUninstallString, SystemComponent and WindowsInstaller to PSADT.Types.InstalledApplication" as an example of this.