- Make callers using `Set-ADTActiveSetup` explicitly specify an ExecutionPolicy for their scripts.
We shouldn't just be doing this as callers may very well be using signed scripts. Just as this is opt-in with PowerShell.exe when running from the commmand line, it needs to be opt-in by the caller also.

For the 3.x translation layer, we just set it back to `Bypass` as it always has been to ensure previous deployments keep trucking along.