- Get rid of `Start-ADTProcess`'s `-NoExitOnProcessFailure` switch entirely.
This was a holdover from the v3.x code and its implementation makes no sense within proper PowerShell standards. The function by default throws a terminating error, and as such, will throw to `Invoke-AppDeployToolkit.ps1`'s catch block and exit out the script anyway. Further to this, every single call to `Start-ADTProcess` within the codebase was supplying the switch to disable it.