- Fix bad typing for for `Open-ADTSession`'s `-DeployAppScriptParameters` parameter.
As per the notes in "- Patch around bad casting in `Deploy-Application.ps1`.", this was initially set up weakly typed due to bad casting in `Deploy-Application.ps1` back in the day. Instead of living with this though, we can just patch around it on the fly to ensure the parameters passed to `Open-ADTSession` are as they should be.