- Rework previous commit to avoid re-defining the scriptblock passed to `Invoke-HKCURegistrySettingsForAllUsers` at all costs.
It's not acceptable that we redefine a caller's scriptblock as it changes the entire error stack and scoping of the code that gets executed within it. We need the code within the scriptblock to execute in the scope and session state where it's defined to ensure a guaranteed, consistent experience for end users.

The changed code draws on combined ideas from Dan G and myself.