- Rework `Import-ADTModuleDataFile` to handle registry overrides for values.
Function and its internal worker have been rewritten to work in reverse. That is, rather than import the caller's config and add missing values from the defaults, use the defaults as a basis and add the caller's values in as overrides. Furthering this, add in the registry values as the final override.