- Resolve `New-ADTTemplate`'s `-Destination` parameter prior to using it.
This is needed for setups like `.\`. Because we read files via `[System.IO.File]::ReadAllLines()` for better performance and encoding handling, we can't do something simple like change the `-Destination` parameter to a  `[System.IO.DirectoryInfo]` object because .NET will resolve this to whatever the working directory was when PowerShell started.