- Add in foundations for classes hive loading within `Invoke-ADTAllUsersRegistryAction`.
By default, we're still not loading it as we need to use P/Invokes to create the necessary symlinks for the Classes mountpoint (reg.exe won't mount into a mountpoint).

Committed mostly because this refactors the function to be nicer and also addresses the issues where there can be a hang when unloading the registry hive while there's still open handles.