- Extract `DeploymentSession.WriteLogEntry()` and move to the LoggingUtilities class.
There's been requests to allow `Write-ADTLogEntry` to be usable without an active session when a specific log file is provided, and that's understandable since the parameters are there. This is also needed for potential async logging to function again so we can address #154.

On the PowerShell side, I tried to use the `[NullString]::Value` singleton as a default value, but found this in my travels: "I know now, that PowerShell will always convert a non-string value (e.g. $null or [NullString]::Value) to an (empty) string, when you use the type modifier [string] for a parameter." (https://stackoverflow.com/questions/51294535/why-does-nullstringvalue-evaluate-differently-with-a-breakpoint). As such, we test the `$PSBoundParameters` dictionary for whether a value is bound or not.