- Don't specify default values within `Invoke-AppDeployToolkit.ps1` for `-DeploymentType` and `-DeployMode`.
Following "- Use `Get-ADTBoundParametersAndDefaultValues` against `Open-ADTSession` so it can include default values.", this was causing the session detection code to not kick in properly as it was considered a specified value.

Rather than some nonsense in the backend, simply remove the default values. If a default value is specified here, it'll be considered authoritative. If none's specified, the DeploymentSession's default values will be used, which is better anyway as it means the default values are now set in one place only.