- Completely rewrite PSADT solution to remove dead code, code that doesn't work, and our own P/Invoke code for CsWin32.
Right now, `Start-ADTProcessAsUser` is completely inoperable (not that it operated right, anyway). New implementation coming in a successive commit.

This commit is also a fine exhibit of "do as I say, not as I do". Given the sheer amount of code we simply were not using and the desire to not only remove that, but also replace the P/Invoke code with CsWin32, the process of stripping it out and re-implementing it bit-by-bit was the most efficient approach.