- Keep all private functions out of the return data from `Get-ADTCommandTable`.
We're doing this as allowing access to the CommandTable also gives out invokable access to our private functions, which is not something we wish to permit at all.

This entire thing had to be done in a bit of an ass-about way due to PowerShell uselessness. We explicitly scope the parameters as Private, however there's no way to retrieve this at runtime. As such, we leverage the `Private:` tag only for compiled modules as it's available in the FunctionDefinitionAst object's name, but for development we rely on file import path to determine the function's privacy level.