- Remove the janky handle `GrantedAccess` checks and do proper thread termination on `NtQueryObject()` hang.
By filtering against GrantedAccess values, we potentially miss out on handles that would otherwise be queryable. The process of setting up the `NtQueryObject()` call in a separate thread is the accepted best solution for this issue, short of writing a driver to do the querying for you. Doing this query on a separate thread is done by:

* psutil: https://github.com/giampaolo/psutil/blob/cedc20a9a750f67adbd4386a6287284f5edd9204/psutil/arch/windows/proc_handles.c#L150-L208
* System Informer: https://github.com/winsiderss/systeminformer/blob/master/phlib/hndlinfo.c#L357-L366

Pavel Yosifovich in his book "Windows Native API Programming" also says: "This works for most object types, except for many file objects - the call to NtQueryObject hangs. This is because many file objects cannot be accessed for name query without grabbing locks that are usually maintained by the file system in question. The only reasonable way I have found to skip “problematic” file objects is to perform the query on a separate thread, and if it blocks for too long, terminate it, and move on. The code becomes more complicated, but cannot be avoided."

The real kicker here - .NET-based threads lost the abort ability in .NET Core, but even then, they likely could not properly about a hung P/Invoke to a native method. As such, we do our own threading directly using unmanaged code.

Our threading is done directly at the kernel level like System Informer as it gives us more control and does not have any reliance on the C Runtime Library (CRT) like I'm lead to believe `CreateThread()` does. Regardless of this, the solution is truly unmanaged - right down to doing this in Shellcode (effectively assembly language for the specific system architectures we support).

The end result is working very well. For now, I've set the timeout on the thread to 125 ms. I've seen older examples time out at 1000 ms, I've seen other projects time out at 100 ms. Before implementing this solution, it took us 250 ms to query all system handles. With the various timeouts in use in mind, I went for 125 ms - half of what it takes, and I like that it's a divisible value of a second.

At the time of commit, an assessment of a debug build of the library yielded a 100% completely clean result on VirusTotal, despite the deep kernel P/Invoke code we're using to get this done.