- Add in necessary code to retrieve high resolution system icons.
The use of System.Drawing.SystemIcons only gives us 32x32 icons. While we're in here porting over the classic UI, let's do it right and get 256x256 jobbies out of the system instead.

Each icon that's pulled is properly cloned so we don't leak GDI handles. They'll stay static in memory for the lifetime of the DLL file, but that's OK with me.