- Quickly get an icon hooked up within the classic CustomDialog.
This required a different approach as reading a HICON straight into a Bitmap doesn't set up the transparency in a way that WinForms required. Proxying through a true Icon object and calling `.ToBitmap()` sets things up correctly.

The remainder of the commit is some slight margin fiddling to make things look good.