- Add new `ProcessToClose` class for sending off to each dialog.
Sending a subset of `RunningProcesses` is erroneous here as it implies that there's only one process to close, when they could be many for the same process name (think Google Chrome and its process per window/tab, etc).

The data used in the UI for displaying processes to close is meant to just be simple data about the name and description of an app that needs closing, irrespective of how many processes that app creates.