- Use CsWin32's raw overload for `WritePrivateProfileString()` so we can properly remove keys or entire sections if required.
The friendly overload for CsWin32 doesn't use nullable strings, therefore it can never be used in accordance with the documentation in that null values for key/value equate to the removal of the section/key, respectively.

Using the raw CsWin32 P/Invoke allows us to resolve the issue. Logged upstream as https://github.com/microsoft/CsWin32/issues/1410.