- Rework INI functions as suggested by @DanGough.
Rightfully so, making `Set-ADTIniValue` do everything breaks the single-use principal. Even though Microsoft do it with their P/Invoke, we don't need to do it here.

As such, make the removal of sections/keys occur in dedicated functions. Further to this, rename the existing functions from `Get-ADTIniValue`/`Set-ADTIniValue` to `Get-ADTIniSectionKeyValue`/`Set-ADTIniSectionKeyValue`. The old functions are there as wrappers, we'll remove them in 4.2.0.