- Revert the "Fixed the window position to prevent bleed on the sides." stuff.
This wasn't the right place to stick such an offset, it should have been done in the above calculation. In saying that, the above calculation already takes care of this and from reviewing the calculated values, the dialog is slammed as hard into whatever corner is chosen as possible. In fact, this made things worse as the dialog jutted out more on the right than before, creating more bleed..