- Correct the work are offsetting within `FluentDialog.PositionWindow()`.
Two things here. One, I was using a bad check to test the dialog position name within the top offsetting. Two, an offset to the left of 1 is too much and -1 is too little. Since these values are doubles, 0.5 is just right. so that there's a consistent amount of offsetting on the x/y planes.