- Rename everything `MessageBox` to `DialogBox`.
Initially I was trying to avoid the word `Dialog` as we use it so prevalently for our main dialogs. The underlying P/Invoke is also called `MessageBox()` so it makes a lot more sense.

However, since our function that calls this is called `Show-ADTDialogBox`, it makes more sense to align the name of everything with it, even if it should be renamed to something like `Show-ADTMessageBox` instead.