- Make new process code not wait for child processes by default.
The old code never did this, and while a large part of writing the new code was for it to do this, it's better that the caller opt into it as required. As @DanGough has pointed out, a number of installers start browsers and other processes and waiting around for them forever in a deployment context is less than ideal as it means the process will never finish executing.