- Fix `lpCompletionCode` testing within `ProcessManager.LaunchAsync()`.
When testing `JOB_OBJECT_MSG_EXIT_PROCESS`, we need to be testing the PID returned via the `lpOverlapped` property. This by rights should be a `NativeOverlapped` object, however C treats this as PVOID and stuffs the PID that sent the message into that pointer. As such, we just take our IntPtr, convert it to an int and read it, instead of just discarding it like we previously were.