- Harden `Open-ADTSession` so any specified parameter can't be null/empty.
This was previously loosey-goosey to allow the frontend to slither through empty strings, etc that we provide in the template. It's not the right approach, rather the frontend should strip those values out before calling `Open-ADTSession`.

We accomplish this by adding the supporting `Remove-ADTHashtableNullOrEmptyValues` function to strip hashtables of key/value pairs where the value is null/empty. We cheekily do this via `Out-String` as it's a cheap way of testing complex objects like arrays, hashtables, and other collection types also.