- Rework `SerializationUtilities.DeserializeFromString()` once more.
Firstly, optimise it down to only pre-cache PSADT-type objects, as well as the recently added primitives. Secondly, it needs to handle array types also, so we define them in a way that's compatible with the serialised XML.

Ideally we wouldn't mutate a ReadOnlyCollection<T> into an Array<T>, however the format of a `ReadOnlyCollection<T>` came through as `<ReadOnlyCollectionOfWindowInfoDRuo7nFw xmlns="http://schemas.datacontract.org/2004/07/System.Collections.ObjectModel" xmlns:i="http://www.w3.org/2001/XMLSchema-instance">`, and I don't know whether the `DRuo7nFw` is a known/reliable constant or not.