- Revert to PSAppDeployToolkit 3.x-era `DeployMode` testing within `DeploymentSession` ctor.
This is incorrect on a technical level, but apparently too many users are out there specifying `-DeployMode Interactive` on the command line, even though Interactive has been, still is, and will likely always be a `DeploymentSession`'s default mode of operation.

There is not a PowerShell command or function out there that I know where an explicit specification of a parameter isn't honoured by it. Not specified? Sure, do whatever; but specified is specified and that's how it should be. I do not agree in any way, shape, or form with the change I am committing here.