- Set up better error handling within the client/server process, and `PSADT.UserInterface.DialogManager.InvokeDialogAction<TResult>()`.
There's a few elements in play here. Firstly, using `Dispatcher.Invoke()` that takes an Action or Delegate does not have a mechanism for handling exceptions. As such, any exception in the dialog code would be unhandled and would crash the entire client/server process. As such, we switch to the generic method and use `Func<TResult>` as the delegate type, and we now receive a thrown error.

Secondly, we had no mechanism in place to send a full-fidelity exception across the pipe. By expanding our serialisation types lookup table with everything the system deems serializable (that is, types that implement `[Serializable]` for the old BinaryFormatter serialisation), and using our dynamic serialisation/deserialisation methods, we're able to serialise a full exception with stack trace across the wire and re-throw it in an inner exception.

Thirdly, as we may be sending ClientException types across the pipe, the definition for ClientException needs to live with the server-side code. The client can still get to it as it depends on the server, however but server cannot/does not depend on the client. We expand this out with a ServerException class as the server-side companion, and use it to re-throw our serialised exception.

Lastly, as we're now more assured that an exception being thrown isn't necessarily bringing down the client/server process, we don't need to terminate the client/server process when a command throws an exception, unless the exception type is `[System.IO.InvalidDataException]`, indicating we received a null or invalid response.