- Get rid of remaining vestiges of old C# naming convention.
The previous convention had complicated file naming to match the namespace/class layout, however this elongates the name by an excessive amount and also severely hampers Visual Studio when it comes to moving classes around namespaces or even just creating new classes as it doesn't know what to name it. The way it is now is the way it should have always been.

The changes here are simply to do with the file names, nothing has changed namespaces or has been relocated elsewhere.