- Revert back to our original user token process code while we further discuss the ConfigMgr "Allow users to view and interact with the program installation" issue.
Regrettably these changes did not address the issue. Probably the only thing that will address the issue would be to run every process via the TokenBroker process (refactored, of course) and then send serialised ProcessResult objects back over a named pipe. This will completely affect the natural flow of process creation and make process cancellation, as well as the passing of anonymous pipe handles, etc through to such processes much harder, if not impossible to perform.