- Make `RequireAdmin` be a parameter against `Open-ADTSession` and not set via the config.
PSADT's most annoying config option is gone. This is not an overarching setting, it's very much something that's per-deployment so it's now settable where it should be. This also removes the exe's attempt to auto-elevate which could be a poor look for a user to receive a UAC credential request on their screen due to misconfiguration. Intune, etc, will already run stuff with the user's privileged token anyway which should handle the requirement.