- Remove unnecessary fixed statements within `Kernel32.WritePrivateProfileString()`.
While CsWin32 doesn't properly declare its method arguments as nullable, they are and therefore we don't need to do this here on our side.

There's only one other P/Invoke where we do fixed pinnings on strings, and that's `ComCtl32.TaskDialog()`. The fixed pinning there is required so that we can appropriately pass through the correct pszIcon `PCWSTR` types, which are not proper strings but rather char pointers to special indices.