- Allow specification of `PROCESS_CREATION_FLAGS.CREATE_BREAKAWAY_FROM_JOB` when creating new process.
Hidden on the PowerShell side, only exposed in `ProcessLaunchInfo` for `ServerInstance` when creating a client/server process. The situation here is that when the calling process is part of a job object, all child processes need to be running in the same session otherwise `CreateProcessAsUser()` can throw an `E_ACCESSDENIED` error. By setting `CREATE_BREAKAWAY_FROM_JOB`, it allows the process to form part of a new job tree to circumvent this problem. It should be noted that this will only work if the parent job object was created with the `JOB_OBJECT_LIMIT_BREAKAWAY_OK` flag to allow child processes to do so.

I believe this will resolve the issues outlined in https://github.com/PSAppDeployToolkit/PSAppDeployToolkit/issues/1506, but it has yet to be tested and confirmed whether this is indeed the case.