- Rework RunAsActiveUser setup and `Get-ADTClientServerUser` to handle running the toolkit as another user from an existing interactive session.
This will just make the client/server process run as the elevated/other user as that's about all we can do. We can probably do something to get the token from the user's explorer.exe process and then use `CreateProcessWithToken()` to start a process since `CreateProcessAsUser()` only works as SYSTEM, however this would not give us stdout/stderr access as it's not available with that API, and stream capture is crucial for the client/server process.