- Fix up some erroneous log lines within `Start-ADTProcess`.
Initially, I wired it up so that `-WindowStyle Hidden` also forced stdout/stderr capture, but it caused problems for certain situations and there could be genuine need to hide the window but not capture stdout/stderr. As such, fix the logging to complete the change from a few months ago.