- Properly check for null within all public ctors for PowerShell safety.
While writing C#, we've got Intellisense to tell us when we might hit a null, but we've got no such thing in the PowerShell side.

This also internalises a number of ctors where only the type needs to be public but we don't expect the caller to construct an object directly.