- Implement the undocumented `EXTENDED_PROCESS_CREATION_FLAG_FORCE_BREAKAWAY` process creation flag.
When we're calling `CreateProcessAsUser()` with a token other than the caller's, if the calling process is part of a job object and the job object does not have the `JOB_OBJECT_LIMIT_BREAKAWAY_OK` flag, we receive an E_ACCESSDENIED` error from the P/Invoke. This isn't a problem for Intune or ConfigMgr, but it certainly is for certain MDMs and RMMs, such as VMware Workspace ONE.

There is an undocumented `PROC_THREAD_ATTRIBUTE` called `PROC_THREAD_ATTRIBUTE_EXTENDED_FLAGS`, which allows us to leverage the also undocumented `EXTENDED_PROCESS_CREATION_FLAG_FORCE_BREAKAWAY` flag, all of which was introduced in Windows 8.1. This flag does what it says on the tin, it allows breakaway no matter how the calling process's job object is set. While note tested with VMware Workspace ONE, it's been tested with `Start-Process -Wait`, which also creates processes attached to a job object without the `JOB_OBJECT_LIMIT_BREAKAWAY_OK` flag.

Lastly, a huge thank you to @jborean93 for pointing me in the right direction. What started as a minor chat ended up turning into a completely working solution for this problem. As we're close to our 4.1.0 release, this last minute change truly will make this release one to remember.