- Commit some code to get the ProcessId of a given `ServiceController` object.
I created this in the process of attempting to create a solution for the problem addressed via "- Implement the undocumented `EXTENDED_PROCESS_CREATION_FLAG_FORCE_BREAKAWAY` process creation flag." that ultimately didn't pan out. It's good code and the functionality could come in handy in the future, so let's not lose it.